Refactor monetary values to use integers instead of floats across the application
- Updated cash movement actions to accept and handle amounts as integers.
- Modified cash register entity and repository methods to use integers for amounts.
- Changed debt payment and debt repository interfaces to use integers for monetary values.
- Adjusted sale repository and service methods to handle totals and amounts as integers.
- Refactored frontend components to format and display currency values using integers.
- Ensured consistent currency formatting across all relevant pages and components.

+++ api/src/Application/Actions/CashRegisters/AddMovementAction.php
@@ -25,7 +25,7 @@ class AddMovementAction
                 $registerId,
                 (int) $user['id'],
                 (string) ($body['type'] ?? ''),
-                (float) ($body['amount'] ?? 0),
+                (int) ($body['amount'] ?? 0),
                 (string) ($body['description'] ?? '')
             );
             $payload = ['success' => true, 'data' => $register];
