+++ api/src/Domain/Repositories/DebtPaymentRepositoryInterface.php
@@ -7,5 +7,5 @@ namespace App\Domain\Repositories;
 interface DebtPaymentRepositoryInterface
 {
     public function findByDebtId(int $debtId): array;
-    public function create(int $debtId, int $userId, ?int $cashRegisterId, float $amount, string $paymentMethod, ?string $notes): int;
+    public function create(int $debtId, int $userId, ?int $cashRegisterId, int $amount, string $paymentMethod, ?string $notes): int;
 }
