+++ api/src/Domain/Repositories/SaleRepositoryInterface.php
@@ -28,15 +28,15 @@ interface SaleRepositoryInterface
         int $cashRegisterId,
         int $userId,
         ?int $customerId,
-        float $total,
+        int $total,
         string $paymentMethod,
-        float $amountPaid,
-        float $changeAmount,
+        int $amountPaid,
+        int $changeAmount,
         ?string $notes
     ): int;
 
     /**
      * Insert a sale item.
      */
-    public function createItem(int $saleId, int $productId, int $quantity, float $unitPrice, float $subtotal): int;
+    public function createItem(int $saleId, int $productId, int $quantity, int $unitPrice, int $subtotal): int;
 }
